add datacard delete test
Former-commit-id: 5831df776925b48f80cc1981b22a967e5b198b3d

+++ opsml/registry/storage/storage_system.py
@@ -399,7 +399,10 @@ class LocalStorageClient(StorageClient):
             read_path:
                 Path to delete
         """
-        return self.client.delete_dir(read_path)
+        if os.path.isdir(read_path):
+            return self.client.delete_dir(read_path)
+
+        return self.client.delete_file(read_path)
 
     @staticmethod
     def validate(storage_backend: str) -> bool:
